+++ Dyalect.Console/Printer.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Dyalect.Library.Json;
 using static System.Console;
 
 namespace Dyalect
@@ -91,7 +92,7 @@ namespace Dyalect
             }
         }
 
-        public static void PrintErrors(IEnumerable<JsonParser.Error> messages)
+        public static void PrintErrors(IEnumerable<JsonError> messages)
         {
             foreach (var m in messages)
                 Error(m.ToString());