+++ Dyalect/Parser/Model/DRange.cs
@@ -11,11 +11,19 @@ namespace Dyalect.Parser.Model
         public DNode From { get; set; }
 
         public DNode To { get; set; }
+        
+        public DNode Step { get; set; }
 
         internal override void ToString(StringBuilder sb)
         {
             From?.ToString(sb);
             sb.Append("..");
+
+            if (Step is not null)
+            {
+                sb.Append('^');
+                Step.ToString(sb);
+            }
             
             if (Exclusive)
                 sb.Append('<');