+++ Docs/Changes.md
@@ -6,6 +6,8 @@
   * Add: _Enhance slicing operator_ ([Issue #371](https://github.com/vorov2/dyalect/issues371)).
   * Add: _Rethink current ranges implementation_ ([Issue #350](https://github.com/vorov2/dyalect/issues/350), [docs](https://github.com/vorov2/dyalect/wiki/Ranges)).
   * Add: _Extend syntax for ranges_ ([Issue #348](https://github.com/vorov2/dyalect/issues347)).
+  * Change: _Dyalect Console updated and now uses JSON parser implementation from Dyalect standard library._
+  * Fix: _Entering ^Z crushes Dyalect console_ ([Issue #416](https://github.com/vorov2/dyalect/issues416)). 
 
 # 0.18.0
   * Change: _Refactoring and redesign of internal compiler infrastructure and foreign module infrastructure._